requirements.txt
blinker==1.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
dnspython==2.1.0
et-xmlfile==1.1.0
eventlet==0.32.0
falcon==3.0.1
Flask==2.0.2
Flask-Login==0.5.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
litex.regon==1.0.6
lml==0.1.0
lxml==4.6.3
MarkupSafe==2.0.1
mongoengine==0.23.1
openpyxl==3.0.9
psycopg2-binary==2.9.1
pyexcel==0.6.7
pyexcel-io==0.6.5
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pymongo==3.12.1
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
texttable==1.6.4
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
xlrd==2.0.1
xlwt==1.3.0
